Fixed issue in which the useEffect hook was continuously making API calls. 'Duckerized' all redux code.
