--- app/client/src/actions/types.js
@@ -1,29 +1,0 @@
-/* 
- Copyright © 2022 Province of British Columbia
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- */
-
-/**
- * Application entry point
- * @author [Jayna Bettesworth](bettesworthjayna@gmail.com)
- * @module
- */
-
-export const GET_COMMUNITIES = 'GET_COMMUNITIES';
-export const ADD_COMMUNITY = 'ADD_COMMUNITY';
-export const DARK_MODE = 'DARK_MODE';
-
-export const GET_POSTS = 'GET_COMMUNITIES';
-export const ADD_POST = 'ADD_COMMUNITY';
-// export const DELETE_COMMUNITY = 'DELETE_COMMUNITY'
