+++ app/client/src/components/communitiesList.js
@@ -22,15 +22,15 @@
 
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
-import { getCommunities } from '../actions/communityActions';
+import { getCommunities } from '../redux/ducks/communityDuck';
 import PropTypes from 'prop-types';
 import Paper from '@mui/material/Paper';
 
 const CommunitiesList = (props) => {
 
   useEffect(() => {
-    props.getCommunities()
-  }, [props])
+    //props.getCommunities()
+  }, [])
 
   return (
     <div>
