+++ app/client/src/components/createCommunity.js
@@ -19,49 +19,31 @@
  * @author [Jayna Bettesworth](bettesworthjayna@gmail.com)
  * @module
  */
- import React, { useEffect, useState } from 'react'
- import jwt_decode from "jwt-decode";
- import { useNavigate } from 'react-router-dom';
- import { useDispatch } from 'react-redux';
- import { createCommunity } from '../actions/communityActions';
- 
- const CreateCommunity = () => {
+import React, { useEffect, useState } from 'react'
+import jwt_decode from "jwt-decode";
+import { useNavigate } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
+import { createCommunity } from '../redux/ducks/communityDuck';
+
+const CreateCommunity = () => {
     const navigate = useNavigate();
     const dispatch = useDispatch();
     const [title, setTitle] = useState('');
     const [description, setDescription] = useState('');
     const [name, setName] = useState('');
-    
+
     async function registerCommunity(event) {
         event.preventDefault();
         const community = {
-            title: title, 
+            title: title,
             description: description,
             creator: name
         };
         dispatch(createCommunity(community));
         navigate('/home')
+    }
 
-    
-        // const response = await fetch('http://localhost:5000/api/Community', {
-        // method: 'POST',
-        // headers: {
-        //   'Content-Type': 'application/json',
-        // },  
-        // body: JSON.stringify({
-        //     title,
-        //     description,
-        //   }),
-        // })
-    
-        // const data = await response.json()
-        // console.log(data);
-        // if (data.status === 'ok'){
-        //  navigate('/home')
-        // }
-      }
-
-      async function userInfo() {
+    async function userInfo() {
 
         const req = await fetch(`${window._env_.API_REF}/profile`, {
             headers: {
@@ -70,49 +52,49 @@
         })
 
         const data = await req.json()
-        if(data.status === 'ok'){
+        if (data.status === 'ok') {
             setName(data.name)
-        }else{
+        } else {
             alert(data.error)
         }
     }
 
-      useEffect(() => {
+    useEffect(() => {
         const token = localStorage.getItem('token')
-        if (token){
+        if (token) {
             const user = jwt_decode(token)
-            if(!user){
+            if (!user) {
                 localStorage.removeItem('token')
                 navigate('/login')
-            }else{
+            } else {
                 userInfo();
-                
+
             }
         }
     }, [])
 
-     return (
-         <div>
-             <h1>Create Community</h1>
+    return (
+        <div>
+            <h1>Create Community</h1>
             <form onSubmit={registerCommunity}>
-               <input 
+                <input
                     value={title}
                     onChange={(e) => setTitle(e.target.value)}
                     type='text'
                     placeholder='Title'
                 />
-                <br/>
-                <textarea 
+                <br />
+                <textarea
                     value={description}
                     onChange={(e) => setDescription(e.target.value)}
                     type='text'
                     placeholder='Description'
                 />
-                <br/>
+                <br />
                 <input type='submit' value='Submit' id='submit' />
             </form>
-         </div>	
-     )
- }
- 
- export default CreateCommunity;
+        </div>
+    )
+}
+
+export default CreateCommunity;
